build: two the gate caught, one of them mine
The sweep runs eight at a time and my fix for a pipe grep closes wrote them all
to one file -- which is the thing CLAUDE.md warns about and the thing that made
a check fail one time in six. Each call has its own name now, beside the
sweep's room rather than in it, because what is in that directory is counted.

And a hole that stopped building when -Wconversion went on: the width it
changes reaches one line under it, so that line comes with the quotation.

+++ src/build.c
@@ -577,7 +577,7 @@ bool kest_bound(KestBuild *build, uint32_t frames, KestLimits *most,
         return false;
     }
     most->heap_bytes = 0;
-    most->call_depth = frames == 0 ? KEST_CALL_DEPTH : frames;
+    most->call_depth = frames == 0 ? KEST_CALL_DEPTH : (uint32_t)frames;
     most->stack_slots = a_chain_of(walked->widest, walked->in_a_turn,
                                    walked->off_the_turns, most->call_depth);
     return true;
